Crate cargo_crev_reviews_wasm

cargo_crev_reviews_wasm

Wasm web app that is the frontend of the application cargo_crev_reviews
repository; version: 2021.925.1120 date: 2021-09-25 authors: Luciano Bestia

This is a member of the workspace. It is the wasm frontend for the GUI application.

Code-flow

Everything is compiled into one single executable binary for Linux: cargo_crev_reviews.
First it opens the default browser with xdg-open on http://127.0.0.1:8182/cargo_crev_reviews/index.html.
If your WSL2 does not have yet a default browser run this:

ln -s "/mnt/c/Program Files/Mozilla Firefox/firefox.exe" /usr/bin/browser_in_win
export BROWSER='/usr/bin/browser_in_win'

The command ln -sf is permanent and persistent. It makes a symbolic link file that stays there forever. But export BROWSER= is NOT persistent. You need to add this command to ~/.bashrc that runs it on every start of terminal.

In the next millisecond the web server starts listening to 127.0.0.1 port 8182.
The first set of requests are GET and response is “static” files embedded in files_mod.rs

browser request for /cargo_crev_reviews/index.html is GET, the response is html text file embedded in files_mod.rs in the function: index_html()
This html is just an empty shell that gets the css and wasm code. There is no real content inside. This concept is Single-page application SPA.

index.html requests: 3 css files, pkg/cargo_crev_reviews.js, pkg/cargo_crev_reviews_bg.wasm, “favicon” icons/icon-032.png. All these requests are GET and responses come from files_mod.rs functions, some are text files and others are base64 files.

the browser imports the wasm module and starts the init function that requests srv_review_list. This responds with: response_method_name, response_html and response_data.

wasm (inside the browser) is rust code. First it matches method_name and calls the appropriate function. It processes the html with the data and inserts it into index.html (the empty shell).

the browser renders our first page. Hooray!

the user click on some button.

the macro on_click! or row_on_click! hides the ugly rust code behind the definition of an event handler in web_sys and calls a function

wasm creates a rpc request and sends/POST to the server

the request is POST, the server first matches the method_name and calls the appropriate function. The function processes the call and prepares some data. It loads the html template.

The response contains the html to be rendered and data to be inserted in this html before rendering.

Modules

auto_generated_mod

cln_methods_review_mod
functions to render the html for reviews

cln_methods_verify_mod

html_mod
generic code to process html

utils_mod
small utility functions

web_sys_mod
helper functions for web_sys, window, document, dom, console, local_storage, session_storage,…

Macros

on_click
Simple macro to set listener of on_click events to an element_id.
fn with 1 arg(element_id): on_click!(element_id, function_ident)

on_keyup
Simple macro to set listener of onkeyup events to an element_id. on_keyup!(“regex_text”, run_regex)

row_on_click
list contains rows, every row item needs a separate event handler the element Id is concatenation of element_prefix plus the row_number

Functions

__wasm_bindgen_generated_wasm_bindgen_start
To start the Wasm application, wasm_bindgen runs this function

wasm_bindgen_start
To start the Wasm application, wasm_bindgen runs this function