Module cargo_crev_reviews_wasm::common_mod

Structs

ReviewFilterData

ReviewItemData

ReviewListData

RpcEmptyData
generic empty data for Rpc

RpcMessageData
generic message for Rpc

RpcRequest
the request_method will be processed on the server

RpcResponse
the response_method will be processed on the client

Enums

RequestMethod
methods available on the server

ResponseMethod
methods available on the client

Functions

find_from
find str from pos_cursor low level panics if pos_cursor is incorrect: Check for bugs in calling functions.

find_pos_after_delimiter
return the position after the delimiter or None Does NOT mutate the pos_cursor, because that is for a higher level logic to decide.

find_pos_before_delimiter
return the position before the delimiter or None Does NOT mutate the pos_cursor, because that is for a higher level logic to decide.

find_range_between_delimiters
find and return the range of the first occurrence between start and end delimiters Success: mutates also the cursor position, so the next find will continue from there Fail: return None if not found and don’t mutate pos_cursor I use type Range to avoid references &str and lifetimes. But the programmer can make the error to apply the range to the wrong vector.

find_range_including_delimiters
find and return the range of the first occurrence including start and end delimiters Success: mutates also the cursor position, so the next find will continue from there Fail: return None if not found and don’t mutate pos_cursor I use type Range to avoid references &str and lifetimes. But the programmer can make the error to apply the range to the wrong vector.

get_delimited_text
returns string between the start end end delimiters without delimiters and the new cursor position