Module cargo_crev_reviews_wasm::html_mod

generic code to process html

Functions

close_modal

decode_5_xml_control_characters

decode 5 xml control characters : “ ’ & < > https://www.liquid-technologies.com/XML/EscapingData.aspx I will ignore all html entities, to keep things simple, because all others characters can be written as utf-8 characters. it is mandatory that text is valid utf-8. https://www.tutorialspoint.com/html5/html5_entities.htm TODO: find a faster method // The standard library replace() function makes allocation,

encode_5_xml_control_characters

encode 5 xml control characters : “ ’ & < > TODO: find a faster method // The standard library replace() function makes allocation,

extract_body_inner

extract_html

inject_into_html

post_request_await_run_response_method

show_modal_html

show_modal_message