Module cargo_crev_reviews_wasm::routing_local_hash_mod

Functions

get_3_url_param_from_hash

get 3 param from hash example “#edit/crate_name/crate_version” -> [“edit”,“crate_name”,“crate_version”] if the param does not exist returns an empty string

router_boilerplate

jump over this boilerplate to router_for_local_hash_routing()

router_for_local_hash_routing

read the url hash parameters for local routing this is a SPA single page application. The page is always index.html then a hash parameter is added for local routing like index.html#edit/crate_name/crate_version the main page is opened only once. It lists the cargo_tree and verify all the dependencies. All other pages are opened in separate tabs. So the user can easily close this tabs and return to the main page. The use of the back button in not recommended.