Module cargo_crev_reviews_wasm::web_sys_mod

helper functions for web_sys, window, document, dom, local_storage, session_storage,…

Functions

close_tab

close the windows/tab in the browser. It works only if it was opened in script wit window.open().

debug_duration

debug write the duration of code execution

fetch_post_response

fetch POST response in Rust with async await for executor spawn_local() return the response as String. Any error will panic.

fetch_response

fetch in Rust with async await for executor spawn_local() return the response as String. Any error will panic.

get_element_by_id

get element by id

get_html_element_by_id

get html element by id

get_input_element_value_string_by_id

get input element value string by id

get_input_html_element_by_id

get input html element by id

get_now_date

get_text

get text from element_id

get_text_area_element_value_string_by_id

get input element value string by id

get_value_of_radio_group_by_name

get elements by name for example radio group

now_performance_millisecond

timestamp with milliseconds

set_inner_html

set inner html to element_id

set_text

set text to element_id

show_snackbar

window

return window object