Crate cargo_crev_reviews_wasm

cargo_crev_reviews_wasm

Wasm web app that is the frontend of the application cargo_crev_reviews
repository; version: 2021.918.1721 date: 2021-09-18 authors: Luciano Bestia

This is a member of the workspace. It is the wasm frontend for the GUI application.

Modules

common_mod
common structs between the backend and frontend WARNING: modify this file only in cargo_crev_reviews project. One automation task will copy it over to cargo_crev_reviews_wasm !

page_review_mod
functions to render the pages for reviews

pages_mod
generic code to process html pages

utils_mod
small utility functions

web_sys_mod
helper functions for web_sys, window, document, dom, console, local_storage, session_storage,…

Macros

on_click
Simple macro to set listener of on_click events to an element_id.
fn with 1 arg(element_id): on_click!(element_id, function_ident)

on_keyup
Simple macro to set listener of onkeyup events to an element_id. on_keyup!(“regex_text”, run_regex)

row_on_click
list contains rows, every row item needs a separate event handler the element Id is concatenation of element_prefix plus the row_number

Functions

__wasm_bindgen_generated_wasm_bindgen_start
To start the Wasm application, wasm_bindgen runs this function

wasm_bindgen_start
To start the Wasm application, wasm_bindgen runs this function