Crate cargo_crev_reviews_wasm

cargo_crev_reviews_wasm

Wasm web app that is the frontend of the application cargo_crev_reviews
repository; version: 2021.1117.1639 date: 2021-11-17 authors: Luciano Bestia

This is a member of the workspace. It is the wasm frontend for the GUI application.
The main member project is cargo_crev_reviews.
Please, continue reading README.md here:

cargo_crev_reviews/README.md

Modules

auto_generated_mod

cln_methods_review_mod
functions to render the html for reviews

cln_methods_verify_mod

cln_methods_version_mod

html_mod
generic code to process html

utils_mod
small utility functions

web_sys_mod
helper functions for web_sys, window, document, dom, local_storage, session_storage,…

Macros

on_click
Simple macro to set listener of on_click events to an element_id.
fn with 1 arg(element_id): on_click!(element_id, function_ident)

on_keyup
Simple macro to set listener of onkeyup events to an element_id. on_keyup!(“regex_text”, run_regex)

row_on_click
list contains rows, every row item needs a separate event handler the element Id is concatenation of element_prefix plus the row_number

url_u
Constructor macro for UrlUtf8EncodedString

Structs

SERVER_FIRST_SUBDIRECTORY

SERVER_HOST
127.0.0.1

SERVER_PORT
8182

UrlPartUtf8Decoded
the url must be utf 8. Only the 5 control characters are encoded.
url has parts or fragments or segments delimited mostly by slash /
every part must be encoded/decoded separately,
to maintain the control character slash /

UrlUtf8EncodedString
Type UrlUtf8EncodedString explicitly informs that the content has been url encoded.
It contains a string with the whole url.
The url is constructed with a special macro, where the dynamic parts are always encoded.
It is impossible to force the developer to properly encode the static part of the url.
But this special type is making this kind of errors difficult, obvious and traceable.

Enums

UrlUtf8Error
thiserror enum

Functions

__wasm_bindgen_generated_wasm_bindgen_start
To start the Wasm application, wasm_bindgen runs this function

get_3_url_param_from_hash
get 3 param from hash example “#edit/crate_name/crate_version” -> [“edit”,“crate_name”,“crate_version”]

wasm_bindgen_start
To start the Wasm application, wasm_bindgen runs this function