Module cargo_crev_reviews_wasm::page_review_mod

functions to render the pages for reviews

Structs

REVIEW_ITEM_DATA
mutable static, because it is hard to pass variables around with on_click events

REVIEW_LIST_DATA

Functions

button_open_crates_io_onclick

button_open_crev_dev_onclick

button_open_lib_rs_onclick

button_open_source_code_onclick

modal_close_on_click

modal_delete

page_review_edit
the code for processing the page_review_edit the data and html are already in static Mutex REVIEW_ITEM_DATA

page_review_error

page_review_list
the code for processing the page rpc_review_list the data is already in static Mutex REVIEW_LIST_DATA

page_review_new

page_review_publish_modal

request_review_delete

request_review_edit_from_list

request_review_list

request_review_new

request_review_new_version

request_review_publish

request_review_save
send rpc requests

request_update_registry_index

store_static_review_list_data
store data in static Mutex because of events like on_click

store_to_review_item_data
store data in static Mutex because of events like on_click